restore iterative consume_point_schedule (cleaner than recursive)
- Revert from recursive back to iterative implementation
- Iterative version is cleaner and more readable with explicit loops
- Fix addition_result_bucket_destinations type from uint64_t to uint32_t
- Eliminates unnecessary type casts throughout
- Keep BB_BENCH profiling markers for performance tracking
- Performance remains identical to recursive version

+++ barretenberg/cpp/src/barretenberg/ecc/scalar_multiplication/scalar_multiplication.hpp
@@ -165,7 +165,7 @@ template <typename Curve> class MSM {
         static constexpr size_t BATCH_OVERFLOW_SIZE = 2;
         std::vector<AffineElement> points_to_add;
         std::vector<BaseField> scalar_scratch_space;
-        std::vector<uint64_t> addition_result_bucket_destinations;
+        std::vector<uint32_t> addition_result_bucket_destinations;
 
         AffineAdditionData() noexcept
             : points_to_add(BATCH_SIZE + BATCH_OVERFLOW_SIZE)
